requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
-e file:///Users/lund/dev/neurosity-sdk-python
    # via -r requirements.in
attrs==23.2.0
    # via
    #   -r requirements.in
    #   cattrs
black==24.3.0
    # via -r requirements.in
cachetools==5.3.3
    # via google-auth
cattrs==23.2.3
    # via -r requirements.in
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
cryptography==42.0.5
    # via jwcrypto
firebase-rest-api==1.11.0
    # via neurosity
google-api-core[grpc]==2.18.0
    # via
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth==2.29.0
    # via
    #   firebase-rest-api
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.16.0
    # via firebase-rest-api
google-cloud-storage==2.16.0
    # via firebase-rest-api
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via google-cloud-storage
googleapis-common-protos==1.63.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.62.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.1
    # via google-api-core
idna==3.6
    # via requests
jwcrypto==1.5.6
    # via python-jwt
mypy==1.9.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.0
    # via black
pathspec==0.12.1
    # via black
pkce==1.0.3
    # via firebase-rest-api
platformdirs==4.2.0
    # via black
proto-plus==1.23.0
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==4.25.3
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
python-dotenv==1.0.1
    # via -r requirements.in
python-jwt==4.1.0
    # via firebase-rest-api
requests==2.31.0
    # via
    #   firebase-rest-api
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
six==1.16.0
    # via firebase-rest-api
typing-extensions==4.11.0
    # via
    #   jwcrypto
    #   mypy
urllib3==2.2.1
    # via requests
